Fix bug with elements with text nodes and child elements
Previously, the extension would set the parent innerHTML to update a child text node. If the parent had other child elements as well as the text node, the child elements would be deleted. This commit fixes that by instead modifying the text nodes nodeValue. This commit also fixes an unrelated issue that caused thousands of mark tags to be generated.